refactor(stats): address review feedback on encoded-resolution fix
Extract readHighestLayerResolution into its own helper file per the
repo's single-responsibility rule (hooks/usePublisherStats already had
several inline helpers), give its reduce() an explicit initial value
(SonarQube flagged the missing one), drop the two comments flagged in
review, and trim the resolution test suite so the edge cases live as
focused unit tests on the extracted helper instead of duplicated
integration tests.
