+++ libs/core/src/hooks/usePublisherStats/helpers/index.ts
@@ -1,0 +1,1 @@
+export { default as readHighestLayerResolution } from './readHighestLayerResolution';
