+++ libs/core/src/hooks/usePublisherStats/usePublisherStats.spec.ts
@@ -84,87 +84,6 @@ describe('usePublisherStats', () => {
 
       expect(stats.resolution.toString()).toBe('640x360');
     });
-
-    it('ignores layers that report no dimensions, which is how idle encodings arrive', async () => {
-      expect.assertions(1);
-
-      const publisher = makePublisher(
-        [
-          makeStatsContainer({
-            video: {
-              layers: [
-                // An encoding that is still active but not transmitting: it keeps reporting a
-                // bitrate and a frame rate, but no dimensions at all.
-                { encodedFrameRate: 4.65, bitrate: 1, scalabilityMode: 'L1T3' },
-                { width: 640, height: 360, encodedFrameRate: 30, bitrate: 800000 },
-              ],
-            },
-          }),
-        ],
-        { videoWidth: 1280, videoHeight: 720 }
-      );
-
-      const { result } = renderHook(() =>
-        usePublisherStats({
-          publisher,
-          publisherStatisticsEnabled: true,
-          fixedFrameRate: 30,
-          queryOptions: { refetchInterval: false },
-        })
-      );
-
-      const stats = await waitForStatsToLoad(result);
-
-      expect(stats.resolution.toString()).toBe('640x360');
-    });
-
-    it('falls back to the captured dimensions when no layer has encoded a frame yet', async () => {
-      expect.assertions(1);
-
-      const publisher = makePublisher(
-        [
-          makeStatsContainer({
-            video: { layers: [{ encodedFrameRate: 0, bitrate: 0 }, { encodedFrameRate: 0 }] },
-          }),
-        ],
-        { videoWidth: 1280, videoHeight: 720 }
-      );
-
-      const { result } = renderHook(() =>
-        usePublisherStats({
-          publisher,
-          publisherStatisticsEnabled: true,
-          fixedFrameRate: 30,
-          queryOptions: { refetchInterval: false },
-        })
-      );
-
-      const stats = await waitForStatsToLoad(result);
-
-      expect(stats.resolution.toString()).toBe('1280x720');
-    });
-
-    it('falls back to the captured dimensions when no layers are reported', async () => {
-      expect.assertions(1);
-
-      const publisher = makePublisher([makeStatsContainer({ video: { layers: [] } })], {
-        videoWidth: 1280,
-        videoHeight: 720,
-      });
-
-      const { result } = renderHook(() =>
-        usePublisherStats({
-          publisher,
-          publisherStatisticsEnabled: true,
-          fixedFrameRate: 30,
-          queryOptions: { refetchInterval: false },
-        })
-      );
-
-      const stats = await waitForStatsToLoad(result);
-
-      expect(stats.resolution.toString()).toBe('1280x720');
-    });
   });
 
   describe('frameRate', () => {
